:sparkles: Change buttons for actions and add them to toolBar
