+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/global/GlobalData.kt
@@ -1,6 +1,8 @@
 package com.github.sloppylopez.moneypennyideaplugin.global
 
+import com.intellij.ui.components.JBTabbedPane
 import javax.swing.JCheckBox
+import javax.swing.JPanel
 
 object GlobalData {
     var downerTabName: Int = 1
@@ -11,4 +13,7 @@ object GlobalData {
     val prompts = mutableMapOf<String, Map<String, List<String>>>()
     val apiKey: String? = System.getenv("OPENAI_API_KEY")
     var explanationButton: JCheckBox? = null
+    var innerPanel: JPanel? = null
+    var nestedPanel: JPanel? = null
+    var tabbedPane: JBTabbedPane? = null
 }
