+++ src/main/kotlin/com/github/sloppylopez/moneypennyideaplugin/helper/ToolWindowHelper.kt
@@ -6,9 +6,6 @@ import com.github.sloppylopez.moneypennyideaplugin.global.GlobalData.tabCounter
 import com.github.sloppylopez.moneypennyideaplugin.services.ProjectService
 import com.github.sloppylopez.moneypennyideaplugin.toolWindow.ButtonTabComponent
 import com.github.sloppylopez.moneypennyideaplugin.toolWindow.MoneyPennyToolWindow
-import com.intellij.openapi.actionSystem.ActionManager
-import com.intellij.openapi.actionSystem.ActionPlaces
-import com.intellij.openapi.actionSystem.DefaultActionGroup
 import com.intellij.openapi.components.service
 import com.intellij.openapi.diagnostic.thisLogger
 import com.intellij.openapi.project.Project
@@ -21,11 +18,10 @@ import com.intellij.ui.content.ContentFactory
 import com.intellij.ui.content.ContentManager
 import java.io.File
 import javax.swing.ImageIcon
-import javax.swing.JComponent
 
 class ToolWindowHelper {
     companion object {
-        private val toolWindowContent = SimpleToolWindowPanel(true)
+        private val toolWindowContent = SimpleToolWindowPanel(false)
         private var moneyPennyToolWindow: MoneyPennyToolWindow? = null
         private var tabbedPane = JBTabbedPane()
 
@@ -58,16 +54,9 @@ class ToolWindowHelper {
                     tabbedPane.setTabComponentAt(i, tabComponent)
                 }
                 toolWindowContent.setContent(tabbedPane)
-                val actionGroup = DefaultActionGroup()
-                actionGroup.add(SendToPromptFileFolderTreeAction(service.getProject()!!))
-                actionGroup.addSeparator()
-                val toolBar = ActionManager.getInstance().createActionToolbar(
-                    "MoneyPennyAI.MainPanel",
-                    actionGroup,
-                    true
-                )
-//                toolBar.targetComponent = toolWindow.component
-                toolWindowContent.toolbar = toolBar.component
+                service.addToolBar(toolWindowContent)
+//                toolWindowContent.componentOrientation = ComponentOrientation.RIGHT_TO_LEFT
+//                service.createToolbarPanel(toolWindowContent)
                 // Add a change listener to handle tab close events
                 addChangeListenerToTabbedPane(tabbedPane, toolWindow.contentManager)
             } catch (e: Exception) {
